scripts/data-generation/requirements.txt
pykafka==2.8.0
cassandra-driver==3.25.0
numpy==1.19.5
tqdm==4.19.5